requirements.txt
cryptography==2.9.2
paho-mqtt==1.5.0
pyperclip==1.8.0